#18 Social Credit initial version implemented.

+++ src/app/controllers/BaseController.java
@@ -40,7 +40,7 @@ public class BaseController extends Controller
     		usr.registrationDate = (Date) a[2];
     		usr.photo = (play.db.jpa.Blob ) a[3];
     		usr.address = (String ) a[4];
-    		usr.balance = 0;
+    		usr.balance = (double) 0;
        		
     		users_new_bee_base.add(usr);
     		
@@ -68,7 +68,7 @@ public class BaseController extends Controller
     				+ " group by e.id");
     		List<Object[]> listRequest = queryRequest.getResultList();
         	
-        	b.balance = listOffer.size() + listRequest.size();
+        	b.balance = (double) (listOffer.size() + listRequest.size());
     		
    			if(count < 6)
    				users_new_bee.add(b);   				
@@ -93,7 +93,7 @@ public class BaseController extends Controller
     		usr.registrationDate = (Date) a[2];
     		usr.photo = (play.db.jpa.Blob ) a[3];
     		usr.address = (String ) a[4];
-    		usr.balance = 0;
+    		usr.balance = (double) 0;
        		
     		users_new_bee_base.add(usr);
        	}  
@@ -114,7 +114,7 @@ public class BaseController extends Controller
     				+ " group by e.id");
     		List<Object[]> listRequest = queryRequest.getResultList();
         	
-        	b.balance = listOffer.size() + listRequest.size();
+        	b.balance = (double) (listOffer.size() + listRequest.size());
     		
    			if(count < 6)
    				users_new_bee.add(b);   				
@@ -139,7 +139,7 @@ public class BaseController extends Controller
     		usr.registrationDate = (Date) a[2];
     		usr.photo = (play.db.jpa.Blob ) a[3];
     		usr.address = (String ) a[4];
-    		usr.balance = 0;
+    		usr.balance = (double) 0;
        		
     		users_new_bee_base.add(usr);
        	}  
@@ -160,7 +160,7 @@ public class BaseController extends Controller
     				+ " group by e.id");
     		List<Object[]> listRequest = queryRequest.getResultList();
         	
-        	b.balance = listOffer.size() + listRequest.size();
+        	b.balance = (double) (listOffer.size() + listRequest.size());
     		
    			if(count < 6)
    				users_new_bee.add(b);   				
@@ -185,7 +185,7 @@ public class BaseController extends Controller
     		usr.registrationDate = (Date) a[2];
     		usr.photo = (play.db.jpa.Blob ) a[3];
     		usr.address = (String ) a[4];
-    		usr.balance = 0;
+    		usr.balance = (double) 0;
        		
     		users_new_bee_base.add(usr);
        	}  
@@ -206,7 +206,7 @@ public class BaseController extends Controller
     				+ " group by e.id");
     		List<Object[]> listRequest = queryRequest.getResultList();
         	
-        	b.balance = listOffer.size() + listRequest.size();
+        	b.balance = (double) (listOffer.size() + listRequest.size());
     		
    			if(count < 6)
    				users_new_bee.add(b);
